Abstracting beam away from laser for use in other weapons, first pass
