+++ Assets/Decontainment/Bot/Weapons/Projectiles/Prefabs/HealBeam.prefab
@@ -1,0 +1,121 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &-4102194359046365534
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2735228162374207918}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f8ce500d1fe32e34fb6b5a87e9f6aed9, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  shotSound: {fileID: 8300000, guid: ad47c1249e426c844af34b69d3aba624, type: 3}
+  maxDistance: 10
+  beamGrowTime: 0.05
+  beamDuration: 0.3
+  beamFadeTime: 0.05
+  healthRegain: 1
+--- !u!1001 &4234311348270267967
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 2249038086161332113, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_Name
+      value: HealBeam
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332115, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_Parameters.colorGradient.key0.r
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332115, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_Parameters.colorGradient.key0.g
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332115, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_Parameters.colorGradient.key0.b
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332115, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_Parameters.colorGradient.key1.r
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332115, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_Parameters.colorGradient.key1.g
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents:
+    - {fileID: 2249038086161332114, guid: 71f2eb08b377f604f8a7bc9aebb20372, type: 3}
+  m_SourcePrefab: {fileID: 100100000, guid: 71f2eb08b377f604f8a7bc9aebb20372, type: 3}
+--- !u!1 &2735228162374207918 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 2249038086161332113, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+    type: 3}
+  m_PrefabInstance: {fileID: 4234311348270267967}
+  m_PrefabAsset: {fileID: 0}
