+++ Assets/Decontainment/Bot/Weapons/Projectiles/Prefabs/HealBeam.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3d6a13d06473b1d47990d13bb983ad6c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
