+++ Assets/Decontainment/Bot/Weapons/Projectiles/Prefabs/LaserBeam.prefab
@@ -1,0 +1,110 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &5186740519727028194
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3802794176953701020}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 8db14bad7549ab84d8a40b8c55f02edd, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  shotSound: {fileID: 8300000, guid: ad47c1249e426c844af34b69d3aba624, type: 3}
+  maxDistance: 10
+  beamGrowTime: 0.05
+  beamDuration: 0.3
+  beamFadeTime: 0.05
+  damage: 1
+--- !u!1001 &3166044964111492365
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: -2263298980897907579, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: beamGrowTime
+      value: 0.05
+      objectReference: {fileID: 0}
+    - target: {fileID: -2263298980897907579, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: beamDuration
+      value: 0.3
+      objectReference: {fileID: 0}
+    - target: {fileID: -2263298980897907579, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: beamFadeTime
+      value: 0.05
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332113, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_Name
+      value: LaserBeam
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_RootOrder
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2249038086161332116, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+        type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 71f2eb08b377f604f8a7bc9aebb20372, type: 3}
+--- !u!1 &3802794176953701020 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 2249038086161332113, guid: 71f2eb08b377f604f8a7bc9aebb20372,
+    type: 3}
+  m_PrefabInstance: {fileID: 3166044964111492365}
+  m_PrefabAsset: {fileID: 0}
