+++ Assets/Decontainment/Bot/Weapons/Projectiles/Prefabs/LaserBeam.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fffbfa5d5f7209e4c818e872f0f7b3e7
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
