+++ Assets/Decontainment/Bot/Weapons/Projectiles/Prefabs/Pulse.prefab
@@ -145,6 +145,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   shotSound: {fileID: 8300000, guid: ad47c1249e426c844af34b69d3aba624, type: 3}
   speed: 8
+  damage: 2
 --- !u!61 &4391388499460265513
 BoxCollider2D:
   m_ObjectHideFlags: 0
