+++ Assets/Decontainment/Bot/Weapons/Projectiles/Scripts/Beam.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eb0b8035048509e4d86c59f73fe35af4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
