+++ Assets/Decontainment/Bot/Weapons/Projectiles/Scripts/HealingBeam.cs
@@ -1,0 +1,12 @@
+﻿using Bot;
+using UnityEngine;
+
+public class HealingBeam : Beam 
+{
+    [SerializeField]
+    private int healthRegain = 1;
+
+    protected override void HealthEffect(Health bh) {
+        bh.HealUp(healthRegain);
+    }
+}
