+++ Assets/Decontainment/Bot/Weapons/Projectiles/Scripts/HealingBeam.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f8ce500d1fe32e34fb6b5a87e9f6aed9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
