+++ Assets/Decontainment/Bot/Weapons/Types/HealWeapon.asset
@@ -10,9 +10,9 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 58acfad9802a29847b6c93bb7747eaf0, type: 3}
-  m_Name: PulseWeapon
+  m_Name: HealWeapon
   m_EditorClassIdentifier: 
-  projectilePrefab: {fileID: 5202299059419044755, guid: 27e0a792cb0a06b4f966045d1ffd499e,
+  projectilePrefab: {fileID: 2735228162374207918, guid: 3d6a13d06473b1d47990d13bb983ad6c,
     type: 3}
   cooldown: 1
-  hardpointColor: {r: 0.313938, g: 0, b: 0.509434, a: 1}
+  hardpointColor: {r: 0, g: 1, b: 0, a: 1}
