+++ Assets/Decontainment/Bot/Weapons/Types/HealWeapon.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a1fcdd241bce75949b5b54439b3e7a8b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
