+++ Assets/Decontainment/Bot/Weapons/Types/LaserWeapon.asset
@@ -12,6 +12,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 58acfad9802a29847b6c93bb7747eaf0, type: 3}
   m_Name: LaserWeapon
   m_EditorClassIdentifier: 
-  projectilePrefab: {fileID: 0}
-  cooldown: 0
-  hardpointColor: {r: 1, g: 1, b: 1, a: 1}
+  projectilePrefab: {fileID: 3802794176953701020, guid: fffbfa5d5f7209e4c818e872f0f7b3e7,
+    type: 3}
+  cooldown: 3
+  hardpointColor: {r: 1, g: 0, b: 0, a: 1}
