+++ Assets/Decontainment/BotData/BotConfigs/HealthShooter.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3e7f9709364c2154ca35a8ac0001145c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
