+++ Assets/Decontainment/BotData/BotConfigs/LaserShooter.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1282de28dcd4b7142aa8bb2184cf545e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
