+++ Assets/Decontainment/BotData/BotConfigs/PulseShooter.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ed60d84064cf82d4794f9fa7c6d7d33c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
